rearranged peptide encoding, fixed unit tests
